package/pyproject.toml
[build-system]
requires = [
    "setuptools>=61.0",
    "wheel",
]
build-backend = "setuptools.build_meta"

[project]
name = "atlas-meshtastic-link"
version = "1.0.0"
description = "Dual-mode Meshtastic link for Atlas Command — gateway and asset in one package."
readme = "README.md"
requires-python = ">=3.10"
license = "MIT"
authors = [
    { name = "ATLAS Team" },
]
keywords = [
    "atlas",
    "meshtastic",
    "link",
    "mesh",
    "gateway",
    "asset",
]
classifiers = [
    "Development Status :: 2 - Pre-Alpha",
    "Intended Audience :: Developers",
    "Programming Language :: Python :: 3",
    "Programming Language :: Python :: 3.10",
    "Programming Language :: Python :: 3.11",
    "Programming Language :: Python :: 3.12",
    "Programming Language :: Python :: 3.13",
    "Topic :: Communications",
    "Topic :: Software Development :: Libraries",
]
dependencies = [
    "atlas-asset-client>=0.3.0",
    "meshtastic>=2.3.0",
    "pypubsub>=4.0.3",
    "pyserial>=3.5",
]

[project.optional-dependencies]
meshtastic = [
    "meshtastic>=2.3.0",
    "pypubsub>=4.0.3",
]
webui = [
    "fastapi>=0.110.0",
    "uvicorn[standard]>=0.29.0",
    "python-multipart>=0.0.9",
]
envelope = [
    "msgpack>=1.0.7",
    "zstandard>=0.22.0",
]
dev = [
    "pytest>=7.0.0",
    "pytest-cov>=4.0.0",
    "pytest-xdist>=3.0.0",
    "fastapi>=0.110.0",
    "httpx>=0.27.0",
    "python-multipart>=0.0.9",
]

[project.urls]
Homepage = "https://github.com/the-Drunken-coder/ATLAS"
Repository = "https://github.com/the-Drunken-coder/ATLAS"

[tool.setuptools.packages.find]
where = [
    "src",
]
include = [
    "atlas_meshtastic_link",
    "atlas_meshtastic_link.*",
]

[tool.setuptools.package-data]
atlas_meshtastic_link = [
    "py.typed",
]
"atlas_meshtastic_link.config.modes" = [
    "*.json",
]

[tool.ruff]
ignore = [
    "E402",
]
